Changes to Statspage and endomondoTasks
so we now have 2 tasks EndomondoWorkouttask and EndomondoAccountTask, the first is just a renamed from EndomondoTask, the second is to grab the persons account first name. I got rid of all the buttons except for the listview clicking those now gives the appropriate graphs, except for duration because i didn't know what to cast the time object to. if your looking to how to tell the data is ready to display wait for the text at the top of the screen to change.